+++ src/routes/(app)/account/+page.svelte
@@ -1,22 +1,22 @@
 <script lang="ts">
-	import { goto } from "$app/navigation";
 	import { page } from "$app/stores";
 	import GithubButton from "$lib/components/GithubButton.svelte";
 	import GoogleButton from "$lib/components/GoogleButton.svelte";
 	import Login from "$lib/components/Login.svelte";
 	import { LinkParameters } from "$lib/types";
-	import { getEncoder } from "$lib/util";
-	import { enhance } from "@sveltejs/kit";
+	import { getEncoder, getDataURL, formFrom } from "$lib/util";
+	import { enhance } from "$app/forms";
 	import { getContext } from "svelte";
 	import type { Writable } from "svelte/store";
-	import type { FormData, PageData } from "./$types";
+	import type { ActionData, PageData } from "./$types";
 	import pen from "$lib/assets/pen.png";
 	import { URLS } from "$lib/addresses";
+	import upload from "$lib/assets/upload.png";
 
-	export let data: PageData;
-	export let form: FormData;
+	export let form: ActionData;
+	const { userData } = $page.data;
 	const stateToken: Writable<string> = getContext("stateToken");
-	const { userData } = data;
+	let pictureDataURL = "";
 
 	let imgUpload: HTMLInputElement;
 	let editMode = false;
@@ -26,25 +26,43 @@
 		referer: $page.url.pathname,
 		stateToken: $stateToken,
 	};
-	const enhanceProps: Parameters<typeof enhance>[1] = {
-		redirect: ({ location }) => goto(location),
-		result: ({ response }) => (form = <any>response),
-		invalid: ({ response }) => (form = <any>response),
-	};
+
+	async function changeEmail() {
+		const email = prompt("New email?");
+		if (!email) return;
+		const res = await fetch($page.url, {
+			method: "POST",
+			body: formFrom({
+				email,
+				passState: getEncoder(LinkParameters).encode({ ...params, method: "email" }),
+			}),
+		}).then((res) => res.json());
+		console.log(res);
+	}
 </script>
 
 {#if !userData}
 	<Login />
 {:else}
 	<div class="flex flex-col gap-3 items-center m-10">
-		<input type="file" bind:this={imgUpload} class="hidden" />
-		<img
-			on:click={() => editMode && imgUpload.click()}
-			alt="profile"
-			class:cursor-pointer={editMode}
-			class="object-cover h-40 w-40 rounded-full border-8 border-white dark:border-neutral-700	"
-			src={userData.picture}
+		<input
+			type="file"
+			on:change={async () => (pictureDataURL = await getDataURL(imgUpload?.files?.[0]))}
+			bind:this={imgUpload}
+			form="editform"
+			name="picture"
+			class="hidden"
 		/>
+		<div
+			id={editMode ? "pcont" : ""}
+			on:click={() => editMode && imgUpload.click()}
+			class:cursor-pointer={editMode}
+			class="relative grid place-items-center h-40 w-40 rounded-full border-8 border-white overflow-hidden dark:border-neutral-700"
+		>
+			<img id="poverlay" alt="upload" class="" src={upload} />
+			<div id="poverlay2" alt="" class="invert dark:invert-0 bg-black/30" />
+			<img alt="profile" class="object-cover w-full h-full" src={pictureDataURL || userData.picture} />
+		</div>
 
 		<div class="grid grid-cols-[auto_auto] gap-x-3">
 			<h2 class="text-right">Name:</h2>
@@ -63,13 +81,13 @@
 					<img on:click={() => (editMode = true)} class="dark:invert cursor-pointer h-6 w-6 inline align-baseline ml-3" src={pen} />
 				</div>
 			{:else}
-				<input form="editform" name="email" class="h-full text-2xl bg-transparent border-b font-bold text-left w-80" value={userData.bio ?? ""} />
+				<input form="editform" name="bio" class="h-full text-2xl bg-transparent border-b font-bold text-left w-80" value={userData.bio ?? ""} />
 			{/if}
 			<h2 class="text-right">Email:</h2>
 			<h2 class="font-bold text-left">{userData.email}</h2>
 			{#if editMode}
 				<div />
-				<form class="contents" id="editform" use:enhance={enhanceProps} method="POST" action="?/editprofile">
+				<form class="contents" id="editform" use:enhance method="POST" action="?/editprofile">
 					<button type="submit" class="cont w-60 mt-3" form="editform">Save</button>
 				</form>
 			{/if}
@@ -80,12 +98,12 @@
 			<div class="bg-lime-300/50 rounded px-1 w-full h-full flex items-center">
 				<span class="text-xl">Email</span>
 			</div>
-			<button class="cont w-full h-full">Change</button>
+			<button on:click={changeEmail} class="cont w-full h-full">Change</button>
 			<div class="{userData.githubID ? 'bg-lime-300/50' : ''} rounded px-1 w-full h-full flex items-center">
 				<span class="text-xl">Github</span>
 			</div>
 			{#if userData.githubID}
-				<form method="POST" use:enhance={enhanceProps} action="?/unlink">
+				<form method="POST" use:enhance action="?/unlink">
 					<button type="submit" name="method" value="githubID" class="cont w-full h-full">Unlink</button>
 				</form>
 			{:else}
@@ -99,7 +117,7 @@
 				<span class="text-xl">Google</span>
 			</div>
 			{#if userData.googleID}
-				<form method="POST" use:enhance={enhanceProps} action="?/unlink">
+				<form method="POST" use:enhance action="?/unlink">
 					<button type="submit" name="method" value="googleID" class="cont w-full h-full">Unlink</button>
 				</form>
 			{:else}
@@ -126,5 +144,25 @@
 	h2,
 	h3 {
 		margin: 0;
+	}
+	#pcont:hover #poverlay,
+	#pcont:hover #poverlay2 {
+		opacity: 1;
+	}
+	#poverlay,
+	#poverlay2 {
+		position: absolute;
+		opacity: 0;
+		transition: 0.5s ease;
+	}
+	#poverlay2 {
+		height: 100%;
+		width: 100%;
+	}
+	#poverlay {
+		height: 50%;
+		width: 50%;
+		z-index: 10;
+		filter: invert(1);
 	}
 </style>