+++ src/routes/(app)/account/link/+server.ts
@@ -1,14 +1,14 @@
 import { URLS } from "$lib/addresses";
 import { IdentificationMethod, LinkParameters } from "$lib/types";
 import { getDecoder } from "$lib/util";
-import { error, type RequestHandler } from "@sveltejs/kit";
+import { error, redirect, type RequestHandler } from "@sveltejs/kit";
 import cookie from "cookie";
 import { z } from "zod";
 import { getTrustedIdentity } from "../lib";
 import { linkExternalAccount } from "./lib";
 
 // we don't need to verify state because this requires login details already
-export const GET: RequestHandler = async ({ url, locals: { sessionToken, userID, state } }) => {
+export const GET: RequestHandler = async ({ url, locals: { sessionToken, userID, state }, cookies }) => {
 	const options = getDecoder(LinkParameters).parse(url.searchParams.get("state"));
 	if (!state || options.stateToken !== state) throw error(400, "invalid state");
 	const { i } = await getTrustedIdentity(url, options.method);
@@ -20,11 +20,6 @@ export const GET: RequestHandler = async ({ url, locals: { sessionToken, userID,
 		sessionToken: z.string().parse(sessionToken),
 	});
 
-	return new Response(null, {
-		status: 300,
-		headers: new Headers({
-			location: URLS.ACCOUNT,
-			"set-cookie": cookie.serialize("state", "", { maxAge: 0 }),
-		}),
-	});
+	cookies.delete("state");
+	throw redirect(300, URLS.ACCOUNT);
 };