+++ src/routes/(app)/account/login/+server.ts
@@ -1,12 +1,12 @@
 import { SESSION_MAXAGE_HOURS } from "$env/static/private";
 import { LoginParameters } from "$lib/types";
 import { getDecoder, log } from "$lib/util";
-import { error, type RequestHandler } from "@sveltejs/kit";
+import { error, redirect, type RequestHandler } from "@sveltejs/kit";
 import cookie from "cookie";
 import { getTrustedIdentity } from "../lib";
 import { loginOrCreateAccount } from "./lib";
 
-export const GET: RequestHandler = async ({ url, locals: { state } }) => {
+export const GET: RequestHandler = async ({ url, locals: { state }, cookies }) => {
 	const options = getDecoder(LoginParameters).parse(url.searchParams.get("state"));
 	log(url.pathname, { state, options });
 	if (!state || options.stateToken !== state) throw error(400, "invalid state");
@@ -15,23 +15,15 @@ export const GET: RequestHandler = async ({ url, locals: { state } }) => {
 	const { userID, newSessionToken } = await loginOrCreateAccount(trustedIdentity, getACD);
 
 	const cookieOpts: cookie.CookieSerializeOptions = {
+		// secure and httponly are default
 		maxAge: options.rememberMe ? +SESSION_MAXAGE_HOURS * 60 * 60 : undefined,
 		path: "/",
-		httpOnly: true,
 		sameSite: "lax",
-		secure: true,
 	};
 
-	const headers = new Headers({
-		location: new URL(options.referer, url).href,
-	});
-	headers.append("set-cookie", cookie.serialize("userID", userID, cookieOpts));
-	headers.append("set-cookie", cookie.serialize("sessionToken", newSessionToken, cookieOpts));
-	headers.append("set-cookie", cookie.serialize("state", "", { maxAge: 0 }));
-	log(url.pathname, "success, setting cookies");
-
-	return new Response(null, {
-		status: 303,
-		headers,
-	});
+	cookies.set("userID", userID, cookieOpts);
+	cookies.set("sessionToken", newSessionToken, cookieOpts);
+	cookies.delete("state", "", { maxAge: 0 });
+	log(url.pathname, "success, set cookies");
+	throw redirect(300, new URL(options.referer, url).href);
 };