+++ src/routes/(app)/restaurants/[slug]/+page.svelte
@@ -48,7 +48,7 @@
 </section>
 
 <style>
-	@import "../../../styles/spinner.module.css";
+	@use "spinner.module.css";
 
 	pre {
 		background-color: rgba(150, 150, 150, 0.1);