+++ src/routes/docs/+layout.svelte
@@ -1,0 +1,6 @@
+<script>
+</script>
+
+<div class="h-screen overflow-auto bg-lime-500/30 p-20">
+	<slot />
+</div>