+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 ## TODO
 
+- [ ] separate page endpoints and server endpoints, to make an api
+  - [ ] api keys for user
+  - [ ] use the passed fetch function
+  - [ ] generate api docs
+- [ ] username field
+- [ ] link for each profile
+- [ ] oispaeliitti integration
+- [x] load user data in layout
+- [x] svelte new cookies api
+- [ ] review code before update
 - [ ] database validation with zod
-- [ ] bio as a new profile field
 - [ ] allow changing profile info and authentication methods, and delete account
 - [ ] view where restaurant owners can see new orders and change their status
 - [ ] secure restaurant add page
@@ -20,6 +29,7 @@
   - [ ] add titles for each page
   - [ ] human readable receipt
   - [ ] info page with link to my website
+- [x] add css loadpath
 - [x] more authentication options
   - [x] show linked identifications on account page
   - [x] GitHub authentication